+++ cli/report.go
@@ -38,11 +38,10 @@ func reportCommand(t *core.Track) *cobra.Command {
 
 	report.PersistentFlags().StringSliceVarP(&options.projects, "projects", "p", []string{}, "Projects to include (comma-separated). All projects if not specified")
 	report.PersistentFlags().StringSliceVarP(&options.tags, "tags", "t", []string{}, "Tags to include (comma-separated). Includes records with any of the given tags")
-	report.PersistentFlags().StringVarP(&options.start, "start", "s", "", "Start date (start at 00:00)")
-	report.PersistentFlags().StringVarP(&options.end, "end", "e", "", "End date (inclusive: end at 24:00)")
 
 	report.AddCommand(timelineReportCommand(t, &options))
 	report.AddCommand(projectsReportCommand(t, &options))
+	report.AddCommand(dayReportCommand(t, &options))
 
 	report.Long += "\n\n" + formatCmdTree(report)
 	return report
@@ -50,7 +49,7 @@ func reportCommand(t *core.Track) *cobra.Command {
 
 func timelineReportCommand(t *core.Track, options *filterOptions) *cobra.Command {
 	timeline := &cobra.Command{
-		Use:     "timeline <days/weeks/months>",
+		Use:     "timeline (days|weeks|months)",
 		Short:   "Timeline reports of time tracking",
 		Aliases: []string{"t"},
 		Args:    util.WrappedArgs(cobra.ExactArgs(1)),
@@ -78,6 +77,9 @@ func timelineReportCommand(t *core.Track, options *filterOptions) *cobra.Command
 			out.Print(timelineFunc(reporter))
 		},
 	}
+	timeline.Flags().StringVarP(&options.start, "start", "s", "", "Start date (start at 00:00)")
+	timeline.Flags().StringVarP(&options.end, "end", "e", "", "End date (inclusive: end at 24:00)")
+
 	return timeline
 }
 
@@ -127,7 +129,66 @@ func projectsReportCommand(t *core.Track, options *filterOptions) *cobra.Command
 			fmt.Print(formatter.FormatTree(tree))
 		},
 	}
+	projects.Flags().StringVarP(&options.start, "start", "s", "", "Start date (start at 00:00)")
+	projects.Flags().StringVarP(&options.end, "end", "e", "", "End date (inclusive: end at 24:00)")
+
 	return projects
+}
+
+func dayReportCommand(t *core.Track, options *filterOptions) *cobra.Command {
+	var blocksPerHour int
+
+	day := &cobra.Command{
+		Use:     "day [DATE]",
+		Short:   "Report of activities over the day",
+		Aliases: []string{"d"},
+		Args:    util.WrappedArgs(cobra.MaximumNArgs(1)),
+		Run: func(cmd *cobra.Command, args []string) {
+			start := util.ToDate(time.Now())
+			var err error
+			if len(args) > 0 {
+				start, err = util.ParseDate(args[0])
+				if err != nil {
+					out.Err("failed to generate report: %s", err)
+					return
+				}
+			}
+			if blocksPerHour <= 0 {
+				out.Err("failed to generate report: argument --width must be > 0")
+				return
+			}
+			filterStart := start.Add(-time.Hour * 24)
+			filterEnd := start.Add(time.Hour * 24)
+
+			filters, err := createFilters(options, false)
+			if err != nil {
+				out.Err("failed to generate report: %s", err)
+				return
+			}
+			filters = append(filters, core.FilterByTime(filterStart, filterEnd))
+
+			reporter, err := core.NewReporter(t, options.projects, filters)
+			if err != nil {
+				out.Err("failed to generate report: %s", err)
+				return
+			}
+			var active string
+			if rec, ok := t.OpenRecord(); ok {
+				active = rec.Project
+			}
+
+			str, err := renderDayTimeline(reporter, active, start, blocksPerHour)
+			if err != nil {
+				out.Err("failed to generate report: %s", err)
+				return
+			}
+			fmt.Print(str)
+		},
+	}
+
+	day.Flags().IntVarP(&blocksPerHour, "width", "w", 3, "Width of the graph, in characters per hour")
+
+	return day
 }
 
 func timelineDays(r *core.Reporter) string {
@@ -205,4 +266,80 @@ func renderTimeline(dates []time.Time, values []float64, unit float64) string {
 	}
 
 	return sb.String()
+}
+
+func renderDayTimeline(reporter *core.Reporter, active string, startDate time.Time, blocksPerHour int) (string, error) {
+	tree, err := core.ToProjectTree(reporter.Projects)
+	if err != nil {
+		return "", err
+	}
+
+	timelines := map[string][]float64{}
+	for pr := range reporter.Projects {
+		timelines[pr] = make([]float64, blocksPerHour*24, blocksPerHour*24)
+	}
+
+	now := time.Now()
+
+	for _, rec := range reporter.Records {
+		endTime := rec.End
+		if rec.End.IsZero() {
+			endTime = now
+		}
+		if endTime.Before(startDate) {
+			continue
+		}
+		start := rec.Start.Sub(startDate).Hours() * float64(blocksPerHour)
+		end := endTime.Sub(startDate).Hours() * float64(blocksPerHour)
+
+		if start < 0 {
+			start = 0
+		}
+		if end > float64(blocksPerHour*24) {
+			end = float64(blocksPerHour * 24)
+		}
+		startIdx := int(start)
+		endIdx := int(end)
+
+		timeline := timelines[rec.Project]
+		for i := startIdx; i <= endIdx; i++ {
+			startProp := start - float64(i)
+			if startProp < 0 {
+				startProp = 0.0
+			}
+			endProp := end - float64(i)
+			if endProp > 1 {
+				endProp = 1.0
+			}
+			timeline[i] += endProp - startProp
+		}
+	}
+
+	timelineStr := map[string][]rune{}
+	for pr, values := range timelines {
+		runes := make([]rune, blocksPerHour*24, blocksPerHour*24)
+		for i, v := range values {
+			runes[i] = util.FloatToBlock(v)
+		}
+		timelineStr[pr] = runes
+	}
+
+	formatter := util.NewTreeFormatter(
+		func(t *core.ProjectNode, indent int) string {
+			fillLen := 16 - (indent + utf8.RuneCountInString(t.Value.Name))
+			var str string
+			if t.Value.Name == active {
+				str = color.BgBlue.Sprintf("%s", t.Value.Name)
+			} else {
+				str = fmt.Sprintf("%s", t.Value.Name)
+			}
+			if fillLen > 0 {
+				str += strings.Repeat(" ", fillLen)
+			}
+			runes, _ := timelineStr[t.Value.Name]
+			return fmt.Sprintf("%s %s", str, string(runes[:]))
+		},
+		2,
+	)
+	return formatter.FormatTree(tree), nil
 }